refactor(intent): add types and re-export from telegram package (GH-804)
Move ConversationMessage, ConversationStore, and AnthropicClient types
to internal/intent package for reuse. Update telegram package to use
type aliases for backward compatibility.

- Add ConversationMessage, ConversationStore, AnthropicClient to internal/intent
- Update telegram/anthropic.go with type aliases pointing to intent package
- Update telegram/conversation.go with type aliases and constructor wrapper
- Move anthropic_test.go and conversation_test.go to internal/intent package
